Move account settungs to separate app
