+++ euth/accounts/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AccountConfig(AppConfig):
+    name = 'euth.accounts'
+    label = 'euth_accounts'
