+++ euth/accounts/forms.py
@@ -1,0 +1,49 @@
+import collections
+
+from django import forms
+from django.utils.translation import ugettext_lazy as _
+
+from euth.contrib import widgets
+from euth.users import models as user_models
+
+
+class ProfileForm(forms.ModelForm):
+
+    class Meta:
+        model = user_models.User
+        fields = ['username', '_avatar', 'description', 'birthdate',
+                  'country', 'city', 'timezone', 'gender', 'languages',
+                  'twitter_handle', 'facebook_handle', 'instagram_handle',
+                  'get_notifications']
+        widgets = {
+            'description': forms.Textarea(),
+            'birthdate': widgets.DateInput(),
+        }
+
+    @property
+    def formsections(self):
+        formsections = collections.OrderedDict([
+            (_('Basic Info'), [
+                'username',
+                '_avatar',
+            ]),
+            (_('Personal Info'), [
+                'description',
+                'birthdate',
+                'country',
+                'city',
+                'timezone',
+                'gender',
+            ]),
+            (_('Ways to connect with you'), [
+                'languages',
+                'twitter_handle',
+                'facebook_handle',
+                'instagram_handle',
+            ]),
+            (_('Notifications'), [
+                'get_notifications',
+            ])
+        ])
+
+        return formsections
