+++ euth/accounts/templates/euth_accounts/base_account.html
@@ -1,0 +1,58 @@
+{% extends "base.html" %}
+{% load i18n thumbnail avatar %}
+
+{% block meta_viewport %}{% endblock %}
+
+{% block content %}
+<div class="container">
+    <div class="dashboard dashboard-bright row">
+        <div class="col-md-3">
+            <div class="dashboard-nav">
+                <div class="dashboard-nav-group">
+                    {# FIXME: unify with new a4dashboard (was dropdown.html) #}
+                    <div class="dropdown">
+                        <button id="dLabel" type="button">
+                            <span class="circled"
+                                  style="background-image: url({% get_avatar request.user 'avatar_small' %});"></span>
+                            <span class="dropdown-title">{{ request.user.username }}</span>
+                        </button>
+                    </div>
+                    <hr>
+
+
+                    <div class="list-group">
+                    {% with request.resolver_match.url_name as url_name %}
+                        <a href="{% url 'account-profile' %}"
+                           class="list-group-item list-group-item-heading {% if url_name == 'account-profile' %}selected{% endif %}">
+                            <i class="fa fa-user" aria-hidden="true"></i> <span>{% trans 'Edit profile' %}</span>
+                        </a>
+                        <a href="{% url 'account_change_password' %}"
+                           class="list-group-item list-group-item-heading {% if url_name == 'account_change_password' %}selected{% endif %}">
+                            <i class="fa fa-lock"></i>
+                            <span>{% trans "Change Password" %}</span>
+                        </a>
+                        <a href="{% url 'account_email' %}"
+                           class="list-group-item list-group-item-heading {% if url_name == 'account_email' %}selected{% endif %}">
+                            <i class="fa fa-envelope" aria-hidden="true"></i>
+                            <span>{% trans 'E-mail Addresses' %}</span>
+                        </a>
+                        <a href="{% url 'socialaccount_connections' %}"
+                           class="list-group-item list-group-item-heading {% if url_name == 'socialaccount_connections' %}selected{% endif %}">
+                            <i class="fa fa-users" aria-hidden="true"></i>
+                            <span>{% trans 'Account Connections' %}</span>
+                        </a>
+                    {% endwith %}
+                    </div>
+                </div>
+            </div>
+        </div>
+        <div class="col-md-9">
+            <div class="dashboard-content">
+                {% block account_content %}{% endblock %}
+            </div>
+        </div>
+    </div>
+</div>
+
+
+{% endblock %}
