+++ euth/accounts/templates/euth_accounts/profile_form.html
@@ -1,8 +1,8 @@
-{% extends "euth_dashboard/base_dashboard.html" %}
+{% extends "euth_accounts/base_account.html" %}
 {% load widget_tweaks i18n form_tags %}
 
 {% block title %}{% trans 'Your user profile' %}{% endblock %}
-{% block dashboard_content %}
+{% block account_content %}
 <h1 class="gray-divider dashboard-title">{% trans 'Your profile' %}</h1>
 
 <div class="general-form">
