+++ euth/dashboard/forms.py
@@ -53,48 +53,6 @@ class CategoryForm(forms.ModelForm):
         exclude = ('module', )
 
 
-class ProfileForm(forms.ModelForm):
-
-    class Meta:
-        model = user_models.User
-        fields = ['username', '_avatar', 'description', 'birthdate',
-                  'country', 'city', 'timezone', 'gender', 'languages',
-                  'twitter_handle', 'facebook_handle', 'instagram_handle',
-                  'get_notifications']
-        widgets = {
-            'description': forms.Textarea(),
-            'birthdate': widgets.DateInput(),
-        }
-
-    @property
-    def formsections(self):
-        formsections = collections.OrderedDict([
-            (_('Basic Info'), [
-                'username',
-                '_avatar',
-            ]),
-            (_('Personal Info'), [
-                'description',
-                'birthdate',
-                'country',
-                'city',
-                'timezone',
-                'gender',
-            ]),
-            (_('Ways to connect with you'), [
-                'languages',
-                'twitter_handle',
-                'facebook_handle',
-                'instagram_handle',
-            ]),
-            (_('Notifications'), [
-                'get_notifications',
-            ])
-        ])
-
-        return formsections
-
-
 class ProjectInviteForm(forms.Form):
     emails = forms.CharField(
         label=_('E-mail addresses of invited users'),
