--- euth/dashboard/templates/euth_dashboard/base_dashboard.html
@@ -1,49 +1,0 @@
-{% extends "base.html" %}
-{% load i18n thumbnail %}
-
-{% block meta_viewport %}{% endblock %}
-
-{% block content %}
-<div class="container">
-    <div class="dashboard dashboard-bright row">
-        <div class="col-md-3">
-            <div class="dashboard-nav">
-                <div class="dashboard-nav-group">
-                    {% include 'euth_dashboard/includes/dropdown.html' with is_user=True title=request.user.username orgs=request.user.organisation_set.all %}
-                    <div class="list-group">
-                        {% url 'dashboard-profile' as dashboard_profile %}
-                        <a href="{{ dashboard_profile }}"
-                           class="list-group-item list-group-item-heading {% if view.menu_item == 'profile' %}selected{% endif %}">
-                            <i class="fa fa-user" aria-hidden="true"></i> <span>{% trans 'Edit profile' %}</span>
-                        </a>
-                        <a href="{% url 'dashboard-password' %}"
-                           class="list-group-item list-group-item-heading {% if view.menu_item == 'password' %}selected{% endif %}">
-                            <i class="fa fa-lock"></i>
-                            <span>{% trans "Change Password" %}</span>
-                        </a>
-                        {% url 'dashboard-email' as dashboard_email %}
-                        <a href="{{ dashboard_email }}"
-                           class="list-group-item list-group-item-heading {% if view.menu_item == 'email' %}selected{% endif %}">
-                            <i class="fa fa-envelope" aria-hidden="true"></i>
-                            <span>{% trans 'E-mail Addresses' %}</span>
-                        </a>
-                        {% url 'dashboard-connections' as dashboard_connections %}
-                        <a href="{{ dashboard_connections }}"
-                           class="list-group-item list-group-item-heading {% if view.menu_item == 'connections' %}selected{% endif %}">
-                            <i class="fa fa-users" aria-hidden="true"></i>
-                            <span>{% trans 'Account Connections' %}</span>
-                        </a>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="col-md-9">
-            <div class="dashboard-content">
-                {% block dashboard_content %}{% endblock %}
-            </div>
-        </div>
-    </div>
-</div>
-
-
-{% endblock %}
