+++ euth_wagtail/templates/account/password_change.html
@@ -1,17 +1,17 @@
-{% extends "euth_dashboard/base_dashboard.html" %}
+{% extends "euth_accounts/base_account.html" %}
 
 {% load i18n %}
 {% load widget_tweaks %}
 
 {% block title %}{% trans "Change Password" %}{% endblock %}
 
-{% block dashboard_content %}
+{% block account_content %}
     <h1 class="gray-divider dashboard-title">{% trans 'Change your password' %}</h1>
 
     <div class="general-form">
         <div class="row">
             <form class="col-md-10" role="form" name="PasswordChangeForm"
-                  action="{% url 'dashboard-password' %}" method="post">
+                  action="{% url 'account_change_password' %}" method="post">
                 {% csrf_token %}
 
                 {% if form.non_field_errors %}
